requirements.txt
gradio
pymongo==3.11
pymongo[srv]==3.11
langchain 
openai 
tiktoken 
googletrans-py
weaviate-client==3.*
pydantic
streamlit==1.30.0
flask
python-dotenv
langchain-core
langchain-mistralai
langchain_community==0.0.13
langchain_openai==0.0.2.post1
qdrant-client
fastembed


